requirements.txt
django~=4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
# wheel==0.41.2
# django-appconf==1.0.5
# django-compressor==4.4
# rcssmin==1.1.1
# rjsmin==1.2.1